Map Generator is working

+++ src/MapGenerator.cpp
@@ -5,7 +5,7 @@
 // Login   <jean_c@epitech.net>
 //
 // Started on  Fri May 15 13:34:23 2015 clément jean
-// Last update Fri May 15 17:56:43 2015 clément jean
+// Last update Fri May 15 18:09:46 2015 polydo_s
 //
 
 #include "MapGenerator.hh"
@@ -46,7 +46,7 @@ void MapGenerator::Show()
 	{
 	  if (i == 0 || i == this->_map.size() - 1 || j == 0 || j == this->_map[i].size() - 1)
 	    {
-	      Wall *w = new Wall(i, j);
+	      Wall *w = new Wall((float)i, (float)j);
 	      this->_map[i][j] = w;
 	    }
 	  if (this->_map[i][j] != NULL)
